Initial configuration setup
This is the initial attempt at setting up configuration management
for the system in a way that will support both configuration updates
from the core system as well as local configuration for individual
sites which should not be overridden.

The setup is based on the Configuration Split module configured with
two settings:

1. Core (for Core configuration) which has file based configuration
2. Local (for individual sites) which has DB based configuration

The intent is that configuration which should be managed by Core is
added to the Core setting during development using complete splits.

By using a wildcard for partial splits for Local we ensure that this
configuration is stored in the database and thus not overriden during
deployment.

The committed configuration is a Minimal installation of Drupal with
the config split module enabled. Configuration for all modules is
currently set to be managed by Core. By using the Minimal we ensure
that we can explicitly control as much as the configuration as
possible.

The only exception is the System module which controls the site name.
We know that this configuration will vary by site and thus explicitly
not adding this configuration to the Core setting proves that the
approach works.