Initial config setup
This is the initial attempt at setting up configuration management
for the system in a way that will support both configuration updates
from the core system as well as local configuration for individual
sites which should not be overridden.

The setup is based on the Config Ignore module.

The committed configuration is a Minimal installation of Drupal with
the config split module enabled. Configuration for all modules is
currently set to be managed by Core. By using the Minimal we ensure
that we can explicitly control as much as the configuration as
possible.

Add documentation and an architecture design record for configuration
management

This should help document how to use the solution and remember why we
chose to go this way on this quite central topic.