Have the php-service depend on a healthy mariadb
By adding a healthcheck to the mariadb docker-compose service and a
depends_on we can ensure a docker-compose up does not allow the php-
service to start before the database is ready.

This protects us from situations where we boot quickly enough to have
a drupal site-install hit a not fully booted mariadb.