Upgrade to PHP 8.0 and Drupal 9.2.7
Based on our coding guidelines we support PHP 8.0 so we might as well
use that for our platform. Update the Lagoon images and CI pipeline
accordingly.

Since Drupal 9.1 does not support PHP 8 we have to update to the
latest version of Drupal, 9.2.7, along with all our other
dependencies as well.

Committed changes in the web dir are a result of updated scaffolded
files.

Using the latest PHP images from Lagoon will also mean that we get
Composer 2.