Output Drupal logs to stdout in JSON format
Outputting logs to stdout will make log output easier to pick up in
an environment with containers.

Outputting logs in JSON format will make logs easier to process with
tools such as Logstash, Loki and cloud platform log processing
services.

Enable and configure the JSONLog module to achieve this. The module
is developed by the Copenhagen Municipality which has similar
requirements as us.

The primary points of the module configuration are:

1. Logging output to stdout
2. Logging everything except debug entries
3. Not truncating log output

2 + 3 should be reasonable since we are either in a local environment
where we are not concerned about resource usage or in the cloud where
we are not immediately concerned about resources.

Consequently they are  excluded by the config ignore list.

JSON replaces the database logging module. We are not interested
in putting strain on the database by writing here anyway.