Use dev snapshot of pa11y-ci
This includes the recent v6 of pa11y which in turn includes the
axe-core v4.3 accessibility engine.

Our accessibility tests for the design system on Storybook uses
axe-core v4.3 so to keep analysis and results relatable it is
worthwhile for us to use an unreleased version of pa11y-ci.

To support installation of unreleased versions we have to authenticate
against GitHub. HTTP auth + token is the easiest way for us to do so.