Add Behat mink debug extension to help debugging
Behat test can be tricky to debug - especially when running on CI
environments. To help in this processes it is especially helpful to
have logs of the DOM and screenshots of the site when failures occur.

MinkDebugExtension is one of many packages offering this. It seems to
be well maintained and compatible with our Chrome Devtools driver.