Add support for loading default content for demonstration purposes
Add the default content used as a foundation for this and a new
DPL example content module which is the module intended for managing
example content in the future.

To demonstrate that the module is working it currently contains
a single node with an attached file. The node refers to a
content type which intentionally is not part of the project
configuration. It should be replaced with proper example content once
the proper content architecture is defined.

You can see the module in action by following these steps:

1. Enable the Field UI module
2. Create a new content type “Article” with default configuration.
3. Add a new File Reference field “File”  with default configuration
   to the “Article” content type.
4. Enable the DPL Example Content module
5. Go to the content overview and see a node called “Example article
   Node” with an attached file “example-file.txt”

Also add the Devel module as it supports the process of configuring
new example content.

Both modules are intentionally not enabled by default.

Add documentation describing usage.