Switch to having dockerfiles in /lagoon
We're switching to a local dev setup that is more similar to what we
use for deployments. First step towards this is to have a lagoon folder
that contains the docker-files we want lagoon to use when we deploy.

For dpl-cms these files will be used when running PR environments