Moving dource.dockerfile to lagoon dir and adding cli local building to build source task.
We need to make some changes in order to make it possible to make pull request environments on the dpl-cms repository.

The model we have chosen so far is a multi-step model that continiously builds new source images
that are being used by other repositories as a part of their docker files.

When working with dpl-cms repository we have only one repository,
so we need to make some changes that are following the example from Amazee:
https://github.com/amazeeio/drupal-example-simple/blob/c13a83cd01f79fad09eb4502d28daeeeda5610d8/docker-compose.yml

When deploying Lagoon parses the docker-compose file and looks at the build properties of the services: cli, php, nginx.
In other sites that dpl-cms (and developer sites) the source image, produced by tagging the dpl-cms repository, is used as a part of the cli images which the other imges inherit.
That is different in a dpl-cms site deploy: The cli image is inherited from lagoon and is running composer install which means the source code is build upon deploy.