Moving source.dockerfile to lagoon dir and adding cli local building to build source task.
We need to make some changes in order to make it possible to make pull request environments on the dpl-cms repository.

The model we have chosen so far is a multi repository model. One repository (dpl-cms) continuously builds new source images and other repositories (environment respositories) are using the source imgaes as a part of their docker files.

When working with and deploying the dpl-cms repository we have chosen to do it differently. In order to work as intended with PR environments following the Lagoon guidelines, we have chosen  only to use one repository.
This is also described in the example from Amazee:
https://github.com/amazeeio/drupal-example-simple/blob/c13a83cd01f79fad09eb4502d28daeeeda5610d8/docker-compose.yml

When deploying Lagoon parses the docker-compose file and looks at the build properties of the services: cli, php, nginx.
In other sites that dpl-cms (and developer sites) the source image, produced by tagging the dpl-cms repository, is used as a part of the cli images which the other imges inherit.
That is different in a dpl-cms site deploy: The cli image is inherited from lagoon and is running composer install which means the source code is build upon deploy.