Fix Adgangsplatformen Behat test
It turns out that Behat steps fail when an exception is thrown - not
if a step returns false. 

Updating this returned in the finding that our oauth/token mock did
not work at all. The request does not have a multipart body. Instead
we do a regular check on the body content.