Rename and refactor source image/registry references
We want to make it possible for a folk of dpl-cms to build and use
custom releases. In an effort to make this process simpler, this
commit does some general renaming of our image and registry references
that makes it a bit simpler to deviate from the default.

The main changes are
1. Move away from the github-centric "organization" and "namespace"
   variables and instead combine these into a single registry reference
2. Reference the image tags we place on release images as "RELEASE_TAG"
   instead of "SOURCE_VERSION"

DDFDPDEL-183