Implement end to end test to show that React components are working
This requires us to actual tokens for both user and library so we
need to mock the library token process as well.

After we have the tokens we can check out the demo page which 
contains a button to add a material to material list and see that
clicking it works i.e. it is shown and sends the expected requests
when loaded and clicked.