Wait for page to be loaded before checking Hello World component
On GitHub Actions the text from the Hello World component is not
available when checking for it. This causes the Behat test to fail.
The logs show that the text is not available in the DOM but the
screenshot actually contain the text. The tests also pass locally.

The error may be caused by timing issues so try to wait a bit to
ensure that the page has loaded before we test.