Fix build error when running CI tests
This change fixes a build error ocurring in our node-chromium image due
to an upstream packaging bug with glibc in Alpine. The bug was triggered
because we were attempting to update all Alpine packages when building
the image. This has been removed as it is already performed upstream,
see https://github.com/uselagoon/lagoon-images/blob/main/images/node-builder/16.Dockerfile