Maintain a up to date an valid version of the Drupal OpenAPI spec
This allows us to use the API with other tools without having a running
site.

Also it allows us to detect changes to the specification which may occur
unintentionally as it partly autogenerated based on configuration.

Our specification it somewhat handheld so ensuring that it is valid is
also helpful.

Node tools swagger-cli and jsome helps us with validation and 
formatting.