Add support for specifying campaign links
The link is where the campaign contents should link to when displayed.

Add the link module and a campaign link field to the campaign node.
The link field is required as a campaign should always have something
to link to. Thus it is placed at the top of the edit form.

Update the campaign endpoint to output the url. This will work
regardless of the link being internal or external.