Initial setup for code generation for FBS client
We use OpenAPI Generator as it seems to produce the most usable client
code from the specification. We also considered
https://jane.readthedocs.io/.

Use v5.2.1 of the generator as it produces a client requiring Guzzle
v6. This is the same version used by Drupal 9. Drupal 10 uses Guzzle 7.
We should upgrade to OpenAPI Generator by then.