Update available facets which can be used for campaign rules
The current ones match the setup from the current Ding2 solution. Going
forward they must match the facets exposed by the new search system.

Also make the default language for the facets in English. This will
allow us to use the configuration translation engine to localize them
in the future.