Implement authentication provider for Adgangsplatformen user tokens
This allows REST resources to require requests to include an 
authorization header with a bearer token which can be validated against
Adgangsplatformen and mapped against a Drupal user.

At the same time we disable CSRF token requirements for such requests.
This puts the CMS at the same level as other business systens.