Configure all REST endpoints to use the dpl_login_user_tokens provider
The frontend will pass a bearer token for the library or user by
default. The UserTokenAuthProvider applies to such requests but for
authentication to work it also has to be a valid authentication provider
for the route. All this logic is handled by
\Drupal\Core\Authentication\AuthenticationManager.

Consequently we configure the campaign and proxy endpoints to use this
instead.

In practice both endpoints are accessible by anonymous users but for
some reason the authentication providers must match anyhow. If not the
resources will return 403 and campaigns will not be displayed.

Update documentation and specification accordingly.

Note that this will not actually print any security data. The OpenAPI 
REST generator module does not support custom authentication providers.
This should not matter much for our use case.