Fix directory name for Varnish Purger
While the module seems to work as is we have been seeing errors as shown
below when trying to install the site. This occurs occasionally both on
CI and sometimes locally. We have not been able to recreate the error
consistently.

The bug is also reported here:
https://www.drupal.org/project/varnish_purge/issues/3346464

VarnishPurgerSettings is a plugin and the error seems to occur because
it has not been detected properly. According to the documentation for
namespace resolution the base namespace for Drupal\varnish_purger\
should be modules/varnish_purger/src/. 
https://www.drupal.org/docs/develop/standards/php/psr-4-namespaces-and-autoloading-in-drupal-8

The module name as defined by the .info.yml file is varnish_purger but
the package name on drupal.org is varnish_purge. This causes the module
to be downloaded to `web/modules/contrib/varnish_purge by default.

With this change the directory for the namespace Drupal\varnish_purger\
becomes modules/contrib/varnish_purger/src/. This brings it in
accordance with the documentation. The contrib directory should not make
a difference.

Update composer.lock after the change.

[...]
Notice: >  [notice] Checked da translation for schemata.
Notice: >  [notice] Checked da translation for varnish_purge.
Notice: >  [notice] Checked da translation for dpl_cms.
Notice: >  [notice] Message: Checked available interface translation updates for 16 projects.
> 
> PHP Fatal error:  Uncaught Drupal\Core\Entity\Exception\NoCorrespondingEntityClassException: The Drupal\varnish_purger\Entity\VarnishPurgerSettings class does not correspond to an entity type. in /app/web/core/lib/Drupal/Core/Entity/EntityTypeRepository.php:98
> Stack trace:
> #0 /app/web/core/lib/Drupal/Core/Entity/EntityBase.php(487): Drupal\Core\Entity\EntityTypeRepository->getEntityTypeFromClass('Drupal\\varnish_...')