Fetch translation files from GitHub pages
The Drupal translation system will only save the status of a translation
if the translation is provided with a timestamp as seen in 
locale_translation_status_save().

When using raw files from GitHub through raw.githubusercontent.com files
are returned without a Last-Modified header. Consequently the timestamp
for these files is set to 0.

Avoid this by pulling translations from GitHub Pages. This serves files
with an appropriate HTTP header.