Replace separate PUML diagrams with inline Mermaid
Now that GitHub has native support for Mermaid.js this makes it easier 
to maintain the diagrams in the future without maintaining a separate 
rendering workflow.

Note that we still have other PUML files in the repository so we cannot
eliminate this entirely, but it should make it easier in the future.