Add cron job for importing translations into Lagoon in the background
Another option might be to change translation.update_interval_days in 
locale.settings.yml as a part of the Drupal configuration. This would
move the translation import process into Drupal Cron handling.

This would support our current usage of automated_cron in Lagoon but it
would not support a very frequent schedule as it is defined in days.