Rename log_args argument to debug from bobheadxi/deployments action
It has been renamed for v1 and the old name causes the the action to 
fail.

Having this info generally available is not relevant anymore so instead
we toggle it based on the runner.debug variable which is set to one if
the developer is running the GitHub Actions job in debug mode.