Adding defaults for dashboard settings
We want to have a common way of adding defaults to both setting forms
and the app renderes.
Since there is dependencies to the general settings, this commit is
touching those as well.
Also we are moving some of the url handling from the dashboard into
separate hooks belonging to the modules that own the routes.