Downgrade Lighthouse CI/lhci to 0.8
Running with lhci 0.9 upgrades to Lighthouse 9.3.0. This reveals new 
problems which causes our tests to fail.

LCHI 0.8 was the original version we used before Dependabot started
suggesting updates. Revert to this to ensure we have a working version.
Then we will update LHCI at a later point in time when we have room to
fix these issues.