Ignore current calls to legacyConfig for React settings
The function has been deprecated to discourage future function calls.
We still have some that we do not want to remove at the moment.

Instead of ignoring all calls we just want to ignore the current ones.

Adding a baseline for known PHPStan errors we do not want to fix at the
moment helps us here so add that.