Update openapi-generator to 7.1.0 to generate PHP 8.1 compatible client
Without this our client will issue deprecation warnings. This has been
fixed in 6.x but we go for the latest version of 7.x to keep things
updated.

See https://github.com/OpenAPITools/openapi-generator/pull/11743

Note that this may cause the current situation to be invalid as the
client depends on Guzzle 7 while Drupaæ 9 still depends on Guzzle 6.

This should be resolved shortly with the update to Drupal 10.