Remove version from DPL libraries
If a version is set then assets will not be regenerated even if contents
change. This makes development more difficoult.
See https://www.drupal.org/project/drupal/issues/3370828

Remove version for now.

We could in theory extract something dynamically from version.json and
modify the libraries but that is for another day.