Configure a new link field for articles for canonical urls
Add a new link field with the following configuration:

- Single value: There can be only one canonical url for a page
- No title: Titles do not make sense in relation to canonical urls
- External only: In the library use case we only work with external
  content as sources.
- Hidden from view modes: Canonical url are meant for metatags not UI

Configure the url from this field to be used as the canonical url
using tokens in metatags.

We do *not* use the field provided by the metatag module as this outputs
way to may different tags which are not relevant to the editors. Having
a single field provides a cleaner experience.

DDFFORM-12