Add state field for events
We base this field on enum_field and the event state enum. This provides
a way to control available states in a reliable way and work with this
way in code easily.

This implies a number of configuration changes:
- Add and enable the enum_field module
-Configure a new enum field for events using the enum

To get human-friendly option names we implement the label() function 
from the enum_field module in our enum.

DDFFORM-23