Replace text field on campaigns
Campaigns currently use the body field to manage contents for text
campaigns. With the addition of WYSIWYG editors these have now gotten
a CKEditor. This makes tests fail as Cypress no longer can enter 
contents into the text field.

In general we do not want a rich text editor for text campaigns.
Consequently we add a new simple text field for storing the text.

Update the campaign endpoint to fetch data from the new field for text
campaigns.

Update the tests to enter content into the new field instead of the
body field.

The old body field is removed accordingly.