Add paragraph for handling event categories and prices
A ticket category consists of two fields:

- Name of the category (simple string)
- Price (decimal with 2 digits)

These are placed on events so we support multiple categories and prices.

We use a simple decimal for prices. Currency fields for Drupal seem to
imply webshops and are closely tied to Drupal Commerce. This is overkill
for our needs. Thus we settle on a decimal instead with a minimum value
of 0,0.

For DKK an integer would probably be sufficient but if we want to take
other currencies into account in the future like EUR fractions will be
required.

As most use cases will use DKK for prices we prefix the field by DKK.
When we get to supporting multiple currencies we will probably have to
alter the prefix with a globally configured currency.

Update Config Ignore accordingly.

For events we ignore all fields on events. We also rearrange fields so
they are alphabetical.

DDFFORM-57