Add brick/math module for working with prices which are decimal numbers
Working with these programmatically is hard to get right.

Use brick/math as that is currently the most popular package for such
things. It is used by many other popular packages which means it may
end up in our stack anyway.

It's version number indicates that it might not be production ready but
the documentation states otherwise.

Alternatives considered:
- https://github.com/PrestaShop/decimal: Not as widely supported or
  updated.
- https://github.com/brick/money: Currently too complex as we do not
  deal with multiple currencies. We may be able to migrate to this as
  it depends on brick/math.
- https://github.com/moneyphp/money: Currently too complex.
  See brick/money.