Add UUID url module
This module provides routes to entities by UUID. Since we have exported
example content by UUID this provides a nice bridge which allows us to
reference urls to the content once it is imported.

This is nice for CI tooling.

As this is only needed for CI we only enable the module during
development and ignore any configuration during export.

DDFFORM-83