Dedupe shared react app data attributes
Although adding load to the rendering process we can cut away a
substantial amount of DOM rendering by removing data attributes from the
other apps that has already been used.

We can do this because the react apps shares the data of some of the
data attributes in a shared store.