Patch Default Content module to overwrite files
When importing content referencing files which already exist in the
filesystem the Default Content module will create new files with the
[filename]_1.[extension] filename and update YAML files with exports
accordingly.

This creates an unnecessary diff when importing content as a part of an
existing site. Patch the module to overwrite any existing files.

The patch applies is not the latest in the issue. This is the original
and simplest change to address the problem:
https://www.drupal.org/project/default_content/issues/3200212#comment-14030757

DDFFORM-83