Added Search_api module along with dependencies. Created view for events.
This commit installs and enables the drupal_search_api module, creates a server, and creates an index over the events based on their date.

Creats a view called 'events' that uses this index to search for events. The only filter that should be enabled is filtering by end_date.

DDFFORM-59