Add new `Files` paragraph + new `Document` Media types
A new `Files` paragraph has been created that makes it possible to add links to uploaded files. This could well have been built into the current `Links` paragraph, but as we would like to keep our paragraphs as simple as possible, we have chosen to do it separately.

A new `Document` Media type that handles PDF files has also been added for now.