Rework more event field templates for event series and instances
Rename existing templates by drop the entity type elements: Node and
event. Now they are going to be used for both nodes and eventseries
so this prefix is unnecessary.

Add more templates for inherited fields on event instances. Inherited
fields are displayed as separate fields with their own ids so we have
to create new templates for these. They can just reuse the existing
field tempaltes.