Refactor rendering of event_description.
The description was being rendered with "render|striptags|trim" to remove wrapper div's. Instead of doing this, i've overwritten the field template used for event description.

In the process i update the element to a div instead of a p, because drupal can't render nested p-tags, which would happen since the field_event_description is a wysiwyg, which gets rendered into a p tag.

Will update the element to div in the design system aswell.

DDFFORM-59