Replace node--list-teaser with eventinstance--list-teaser.
Since we are changing the view from using node events to using eventInstance entity, the view mode is changed.

I've also refactored the twig code a bit, to handle more of the logic at the top for the file, to increase readability in the template itself.

Please have a look at this refactored logic and see if you agree.

DDFFORM-118