Limit configuration diffing to core configuration
We have seen from experience that language differences acting weirdly
where imports do not seem to be entirely completed. This results in
language changes in CI which we cannot replicate locally.

To address this we have tried the following:

- Digging through issues queues. The following issues have not provided
  any answers yet:
  - https://www.drupal.org/project/drupal/issues/3103440
  - https://www.drupal.org/project/drupal/issues/3221375
  - https://github.com/drush-ops/drush/issues/4610
- Trying out https://github.com/zaporylie/drush-current-language

Running drush config-import twice seems make the problem go away but we
do not want to go there yet.

To address the situation we limit changes to the default language 
configuration. Checking the configuration here is at least better than
what we have now.