Refactor twig logic into preprocess.
Twig template was for eventisntance--list-teaser was containing too much logic for handling fields values for the view mode.

This logic has been moved into a preprocess for the view mode on eventinstance entities in order to clean that up.

Currently, there is an issue with an incorrect timezone being applied to dates being show on the website.

In this preprocess i've added ` $site_timezone = date_default_timezone_get();` which is i use to get the timezone from the site, and format the dates.
Please comment on this approach.

DDFFORM-118