Change title level for cards
Based on the observation that cards are shown as a part of a larger
element e.g. a slider which will have its own headline, and this
element will be used as a paragraph on a page it seems reasonable to
assume that headlines on cards should be level 3.

Our styling is based on the CSS class anyway so changing the HTML
element should not make a difference visually.