'Card grid automatic' paragraph, to embed a filtered view. DDFFORM-127
This is an addition to the 'card grid manual' paragraph, where the
editor chooses 1-6 contents. This paragraph allows the editor to
choose a content type, to display the latest 6 items.
In the future, we will also introduce a tagging system.

Ontop of that, the view and preprocess functionality has been
made general enough to easily add other view styles and filters.