Patch Dynamic Entity Reference module to clarify field category
Current the Dynamic Entity Reference module uses the "reference"
category. This causes confusion as other fields normally use the
"Reference" category. If the Dynamic Entity Reference field is chosen
by accident this can cause problems. One example is that this field
type is not supported by Field Inheritance.

Patch the module so this does not happen.

The patch has already been merged into the 3.x branch but not released
so we should be able to drop the patch once the next version is
released.
https://www.drupal.org/node/3405253