Add 'cancel' button to event-series-date-confirm flow. DDFFORM-284
When saving a series, and changing the dates, the recurring_events module
will make a "custom" submit button, showing a 'are you sure' confirm.
However, this does not include a 'cancel' button. This is because cancel
basically just means navigating away, but, this is confusing for editors.
We'll add our own "button", that is really just a link to the content
overview page.