Disable purge late runtime processor
We have seen site installations fail with the following error message:

The Drupal\varnish_purger\Entity\VarnishPurgerSettings class does not
correspond to an entity type.

The stacktrace in the following issue indicates that the error occurs
as a part of the LateRuntimeProcessor
https://www.drupal.org/project/varnish_purge/issues/3346464

Try to disable the processor to see if this addresses the problem.

Lagoon suggests using the so this might not be a permanent solution
but we can try.
https://docs.lagoon.sh/applications/drupal/services/varnish/#usage-of-purge-late-runtime-processor