Patch Purge module to try to combat NoCorrespondingEntityClassException
We have seen problems loading Varnish Purger configuration entities
during some site installs on CI. This results in an error when running
task ci:reset:

The Drupal\varnish_purger\Entity\VarnishPurgerSettings class does not
correspond to an entity type.

This occurs when running drush cache:rebuild -y immediately after
running drush site-install --existing-config -y.

We cannot replicate this locally.

History shows that is may be related to the LateRuntimeProcessor
https://www.drupal.org/project/varnish_purge/issues/3346464

We have tried to address this by changing the order of the steps,
disabling the LateRuntimeProcessor but without any success.

We believe this may be temporary so try to combat these by logging the
problem and continuing the process.

This patch works in the same area as an already merged but unreleased
change so include this patch as well and base the current patch on
this. This should make it easier to move forward with our own change
in the future.