Add new custom module `Footer`
This module introduces a custom form for appending content to the footer. It features a multi-value form with a title and its content. The content can be utilized for displaying various elements such as brief information, addresses, and importantly, links.

I have integrated a third-party module `multivalue_form_element` to facilitate the ability to reorder items.

In the `submitForm` method, I have implemented code that discards items lacking content. Currently, this is the sole method to delete items.