Add concept of unregistered user tokens
The UnregisteredUserTokensProvider class extends the UserTokensProvider
so they share functionality. The CRUD operations are the same.

A delete method is added in order to be able to deleta a token again. We
need that in order to make sure that we can delete a unregistered user
token when a normal user token is "taking over".

The user-token endpoint is changed so it will expose an uregistered
token instead of a user token if present.