Fix display of recommendation apps on Lagoon
Currently materials are not shown when editors add a recommendation
paragraph and provide a custom title and description. This occurs on
Lagoon environments but not locally. No React app is rendered at the
expected location at {{ react_app }} in
paragraph--recommendation--recommended-material.html.twig.

It appears as if dpl_paragraphs_preprocess_paragraph__recommendation()
is not run here.

To address this the handling of recommendations is refactored to
remove the additional theme suggestion and instead implement the
variation directly in the paragraph--recommendation.html.twig
template.

Tests on Lagoon shows that this works.

I do not have an explanation for why this is the case or why it works
locally but not on Lagoon. I have tried the following things to
recreate the issue locally but without success:

- Bypassing Varnish and accessing the page through NGINX
- Changing the LAGOON_ENVIRONMENT_TYPE to not disable caching