Patch dynamic_entity_reference to work with entity_clone.
According to the issue, this is really an problem with Drupal core,
but I have tried to introduce the Core patch instead, and it still
does not work.
Without this patch, cloning fails with a WSOD, whenever you try to
clone a peice of content with dynamic_entity_reference field.