Reenable config ignore
To allow local administrators to modify configuration and avoid this
from being overridden we use the config_ignore module.

As we do not need to handle configuration from arbitrary modules yet
we can use config_ignore with a list of configuration entries that
should be ignored during config import.

These settings are a result of:
- Logging in as a local editor
- Updating all the elements under/admin/config that this role has
  access to + OpenID Adgangsplatformen config
- Exporting configuration
- Copy/pasting the configuration entries into the Config Ignore
  settings
- Exporting configuration again