Add patch for entity_clone, that allows redirect to edit. DDFFORM-373
When cloning / using a entity template, the user is redirected
to the view of the content.
This is pretty confusing, especially in the context of the using
templates to create new content.
I've found a patch that adds it as a module option, to set the
redirect to the edit page instead of canoncial.

https://www.drupal.org/project/entity_clone/issues/2708731#comment-13224285