Enable locking for all content types, event series and instances
These are the primary elements we expect editors to work with.

We could consider enabling this for entities that editors might also
work with individually like media, taxonomy and menu links but we
do not expect editors to update these frequently nor for editing of
these to take a long time.

Here locking might not be necessary.