WIP: Add dynamic toggle for alt-text on image
This commit introduces a toggle feature to hide and remove the required attribute from the alt text on images. The functionality operates as expected when editing an already uploaded image. However, there are issues when adding a media paragraph and uploading a new image, as I am unable to hook into the form correctly. I attempted to use media-library-add-form-upload, but it was unsuccessful.

*Note: This task was time-boxed, and no additional time could be allocated.